Moved commands/root module on more appropriate cli/root
